+++ src/compas_ghpython/artists/circleartist.py
@@ -3,25 +3,21 @@ from __future__ import absolute_import
 from __future__ import division
 
 import compas_ghpython
-from compas_ghpython.artists._primitiveartist import PrimitiveArtist
+from compas.artists import PrimitiveArtist
+from .artist import GHArtist
 
 
-__all__ = ['CircleArtist']
-
-
-class CircleArtist(PrimitiveArtist):
+class CircleArtist(GHArtist, PrimitiveArtist):
     """Artist for drawing circles.
 
     Parameters
     ----------
-    primitive : :class:`compas.geometry.Circle`
+    circle : :class:`compas.geometry.Circle`
         A COMPAS circle.
-
-    Other Parameters
-    ----------------
-    See :class:`compas_ghpython.artists.PrimitiveArtist` for all other parameters.
-
     """
+
+    def __init__(self, circle, **kwargs):
+        super(CircleArtist, self).__init__(primitive=circle, **kwargs)
 
     def draw(self):
         """Draw the circle.
@@ -29,14 +25,13 @@ class CircleArtist(PrimitiveArtist):
         Returns
         -------
         :class:`Rhino.Geometry.Circle`
-
         """
-        circles = [self._get_args(self.primitive)]
+        circles = [self._get_args(self.primitive, self.color)]
         return compas_ghpython.draw_circles(circles)[0]
 
     @staticmethod
-    def _get_args(primitive):
+    def _get_args(primitive, color=None):
         point = list(primitive.plane.point)
         normal = list(primitive.plane.normal)
         radius = primitive.radius
-        return {'plane': [point, normal], 'radius': radius, 'color': None, 'name': primitive.name}
+        return {'plane': [point, normal], 'radius': radius, 'color': color, 'name': primitive.name}